Change workspace click event. Add ability to move a cell. And more.
Split the workspaceMouseClick() function into the more precise workspaceMouseDown() and workspaceMouseUp() functions. Move a cell and its dendrites when the user drags it, checking for collision with other cells as appropriate. Add Cell::oldX and Cell::oldY properties to keep track of cell movements (so they can be undone). Remove graphMidpointX and drawingDendrite global variables. Add ignoreCellID parameter to checkForCollision() and checkForRoom() functions.